# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Радар вузов API
  description: '

    CRM контроля взаимодействия ИТ Школы Ростелекома с вузами.


    Все методы `/api/v1` требуют `Authorization: Bearer <access_token>` из Keycloak.

    Ошибки возвращаются в формате `application/problem+json` с полем `code`

    из каталога в ARCHITECTURE.md.

    '
  version: 0.1.0
paths:
...
  /api/v1/reports:
    get:
      tags:
      - reports
      summary: Свои задания на отчёт
      operationId: read_reports_api_v1_reports_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ReportJobOut'
                type: array
                title: Response Read Reports Api V1 Reports Get
        '401':
          description: AUTH_REQUIRED
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      security:
      - HTTPBearer: []
    post:
      tags:
      - reports
      summary: Заказать отчёт
      description: 'Отчёт строится в фоне: ответ содержит id задания, дальше состояние смотрят в `GET
        /api/v1/reports/{id}`. Колонки: group, counterparty, university, direction, program, product,
        stage, owner, contract, license_valid_until, days_on_stage, signals.'
      operationId: post_report_api_v1_reports_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportCreate'
        required: true
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportJobOut'
        '401':
          description: AUTH_REQUIRED
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: VALIDATION_ERROR · REPORT_TOO_LARGE
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      security:
      - HTTPBearer: []
  /api/v1/reports/{report_id}:
    get:
      tags:
      - reports
      summary: Состояние отчёта
      operationId: read_report_api_v1_reports__report_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: report_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Report Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportJobOut'
        '401':
          description: AUTH_REQUIRED
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: NOT_FOUND
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: Validation Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /api/v1/reports/{report_id}/file:
    get:
      tags:
      - reports
      summary: Скачать готовый отчёт
      operationId: read_report_file_api_v1_reports__report_id__file_get
      security:
      - HTTPBearer: []
      parameters:
      - name: report_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Report Id
      responses:
        '200':
          description: Файл отчёта
          content:
            application/octet-stream: {}
        '401':
          description: AUTH_REQUIRED
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: NOT_FOUND
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '304':
          description: С прошлого запроса ничего не изменилось
        '422':
          description: Validation Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
...
components:
  schemas:
...
    ErrorCode:
      type: string
      enum:
      - VALIDATION_ERROR
      - AUTH_REQUIRED
      - AUTH_FORBIDDEN
      - NOT_FOUND
      - INTERACTION_VERSION_CONFLICT
      - INTERACTION_DUPLICATE
      - WF_TRANSITION_NOT_ALLOWED
      - WF_COMMENT_REQUIRED
      - WF_ATTACHMENT_REQUIRED
      - WF_VERSION_NOT_DRAFT
      - WF_MIGRATION_MAP_INCOMPLETE
      - FILE_TYPE_NOT_ALLOWED
      - FILE_TOO_LARGE
      - IMPORT_MAPPING_INVALID
      - REPORT_TOO_LARGE
      - INTEGRATION_UNAVAILABLE
      - INTERNAL_ERROR
      title: ErrorCode
...
    FieldError:
      properties:
        field:
          type: string
          title: Field
        message:
          type: string
          title: Message
      type: object
      required:
      - field
      - message
      title: FieldError
...
    Problem:
      properties:
        type:
          type: string
          title: Type
          default: about:blank
        title:
          type: string
          title: Title
        status:
          type: integer
          title: Status
        code:
          $ref: '#/components/schemas/ErrorCode'
        detail:
          anyOf:
          - type: string
          - type: 'null'
          title: Detail
        trace_id:
          type: string
          title: Trace Id
        errors:
          items:
            $ref: '#/components/schemas/FieldError'
          type: array
          title: Errors
          default: []
      type: object
      required:
      - title
      - status
      - code
      - trace_id
      title: Problem
      description: Тело ответа с ошибкой.
...
    ReportCreate:
      properties:
        format:
          type: string
          enum:
          - xlsx
          - xls
          - csv
          - pdf
          - json
          title: Format
          default: xlsx
        encoding:
          type: string
          enum:
          - utf-8
          - windows-1251
          title: Encoding
          description: 'Кодировка CSV: utf-8 (с BOM, без потерь) или windows-1251 для старых программ'
          default: utf-8
        period_from:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Period From
        period_to:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Period To
        group_id:
          items:
            type: string
            format: uuid
          type: array
          title: Group Id
        university_id:
          items:
            type: string
            format: uuid
          type: array
          title: University Id
        direction_id:
          items:
            type: string
            format: uuid
          type: array
          title: Direction Id
        program_id:
          items:
            type: string
            format: uuid
          type: array
          title: Program Id
        product_id:
          items:
            type: string
            format: uuid
          type: array
          title: Product Id
        owner_id:
          items:
            type: string
            format: uuid
          type: array
          title: Owner Id
        stage_code:
          items:
            type: string
          type: array
          title: Stage Code
        status:
          items:
            type: string
            enum:
            - active
            - paused
            - completed
            - cancelled
          type: array
          title: Status
          description: Состояние записи; без фильтра отменённые не попадают
        search:
          anyOf:
          - type: string
            maxLength: 100
          - type: 'null'
          title: Search
        columns:
          items:
            type: string
          type: array
          minItems: 1
          title: Columns
          description: 'Из набора: group, counterparty, university, direction, program, product, stage,
            owner, contract, license_valid_until, days_on_stage, signals'
      type: object
      title: ReportCreate
      description: Те же фильтры, что у списка взаимодействий, плюс колонки и формат файла.
    ReportJobOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        status:
          type: string
          title: Status
          description: queued, running, done или failed
        progress:
          type: integer
          title: Progress
          description: Процент готовности
        format:
          type: string
          title: Format
        row_count:
          anyOf:
          - type: integer
          - type: 'null'
          title: Row Count
        error_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Code
          description: Код ошибки из каталога, если отчёт не построен
        created_at:
          type: string
          format: date-time
          title: Created At
        finished_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Finished At
      type: object
      required:
      - id
      - status
      - progress
      - format
      - row_count
      - error_code
      - created_at
      - finished_at
      title: ReportJobOut
...
  securitySchemes:
    HTTPBearer:
      type: http
      description: Access token из Keycloak
      scheme: bearer